[refactor] NavGraph 코드 별도 파일 및 함수로 분리
